requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
arrow==0.7.0
itsdangerous==0.24
pymongo==3.0.3
python-dateutil==2.4.2
six==1.10.0
